+++ projects/ng-date-picker/src/lib/data/default-date-options.ts
@@ -5,76 +5,17 @@
  */
 import { DEFAULT_DATE_OPTION_ENUM } from '../constant/date-filter-enum';
 import { ISelectDateOption } from '../model/select-date-option';
+import { createOption } from '../utils/date-picker-utilities';
 
 export const DEFAULT_DATE_OPTIONS: ISelectDateOption[] = <ISelectDateOption[]>[
-  {
-    optionLabel: 'Today',
-    optionKey: DEFAULT_DATE_OPTION_ENUM.DATE_DIFF,
-    dateDiff: 0,
-    isSelected: false,
-    isVisible: true,
-  },
-  {
-    optionLabel: 'Yesterday',
-    optionKey: DEFAULT_DATE_OPTION_ENUM.DATE_DIFF,
-    dateDiff: -1,
-    isSelected: false,
-    isVisible: true,
-  },
-  {
-    optionLabel: 'Last 7 Days',
-    optionKey: DEFAULT_DATE_OPTION_ENUM.DATE_DIFF,
-    dateDiff: -7,
-    isSelected: false,
-    isVisible: true,
-  },
-  {
-    optionLabel: 'Last 30 Days',
-    optionKey: DEFAULT_DATE_OPTION_ENUM.DATE_DIFF,
-    dateDiff: -30,
-    isSelected: false,
-    isVisible: true,
-  },
-  {
-    optionLabel: 'Last Month',
-    optionKey: DEFAULT_DATE_OPTION_ENUM.LAST_MONTH,
-    dateDiff: 0,
-    isSelected: false,
-    isVisible: true,
-  },
-  {
-    optionLabel: 'This Month',
-    optionKey: DEFAULT_DATE_OPTION_ENUM.THIS_MONTH,
-    dateDiff: 0,
-    isSelected: false,
-    isVisible: true,
-  },
-  {
-    optionLabel: 'Month To Date',
-    optionKey: DEFAULT_DATE_OPTION_ENUM.MONTH_TO_DATE,
-    dateDiff: 0,
-    isSelected: false,
-    isVisible: true,
-  },
-  {
-    optionLabel: 'Week To Date',
-    optionKey: DEFAULT_DATE_OPTION_ENUM.WEEK_TO_DATE,
-    dateDiff: 0,
-    isSelected: false,
-    isVisible: false,
-  },
-  {
-    optionLabel: 'Year To Date',
-    optionKey: DEFAULT_DATE_OPTION_ENUM.YEAR_TO_DATE,
-    dateDiff: 0,
-    isSelected: false,
-    isVisible: true,
-  },
-  {
-    optionLabel: 'Custom Range',
-    optionKey: DEFAULT_DATE_OPTION_ENUM.CUSTOM,
-    dateDiff: 0,
-    isSelected: false,
-    isVisible: true,
-  },
+  createOption('Today', DEFAULT_DATE_OPTION_ENUM.DATE_DIFF, 0),
+  createOption('Yesterday', DEFAULT_DATE_OPTION_ENUM.DATE_DIFF, -1),
+  createOption('Last 7 Days', DEFAULT_DATE_OPTION_ENUM.DATE_DIFF, -7),
+  createOption('Last 30 Days', DEFAULT_DATE_OPTION_ENUM.DATE_DIFF, -30),
+  createOption('Last Month', DEFAULT_DATE_OPTION_ENUM.LAST_MONTH),
+  createOption('This Month', DEFAULT_DATE_OPTION_ENUM.THIS_MONTH),
+  createOption('Month To Date', DEFAULT_DATE_OPTION_ENUM.MONTH_TO_DATE),
+  createOption('Week To Date', DEFAULT_DATE_OPTION_ENUM.WEEK_TO_DATE, 0, false),
+  createOption('Year To Date', DEFAULT_DATE_OPTION_ENUM.YEAR_TO_DATE),
+  createOption('Custom Range', DEFAULT_DATE_OPTION_ENUM.CUSTOM),
 ];