+++ projects/ng-date-picker/src/lib/model/select-date-option.ts
@@ -23,7 +23,7 @@ export interface ISelectDateOption {
 
   // boolean value used to show or hide label in list.
   isVisible: boolean;
-  callBackFunction: () => DateRange<Date>;
+  callBackFunction?: () => DateRange<Date>;
 }
 
 export class SelectDateOption implements ISelectDateOption {
@@ -32,5 +32,5 @@ export class SelectDateOption implements ISelectDateOption {
   dateDiff = 0;
   isSelected = false;
   isVisible = false;
-  callBackFunction!: () =>  DateRange<Date>;
+  callBackFunction!: () => DateRange<Date>;
 }