+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 # Changelog
+## [2.3.0] - 2025-09-04
+- Added Signals for state changes.
+- Fist view not showing start date issue fixed.
+
 ## [2.2.9] - 2025-08-21
 - Added support for displaying the selected date range in a specified date format on the calendar.  
 - Added a button to for suffix icons.